% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bradypus}
\alias{bradypus}
\title{Occurrence records and background data for the brown-throated three-toed sloth, Bradypus variegatus}
\format{An object of class \code{data.frame} with 1116 rows and 15 columns.}
\usage{
bradypus
}
\description{
A dataset containing environmental data at 116 Bradypus variegatus occurrence points
and 1000 background points in South and Central America.  Occurrence data are from
Anderson and Handley (2001); see Phillips et al. (2006) for descriptions of the 
predictor variables.
}
\references{
Anderson, R. P. and Handley, Jr., C. O. (2001). A new species of three-toed sloth (Mammalia: Xenarthra) from Panama, with a review of the genus Bradypus. Proceedings of the Biological Society of Washington 114, 1-33.

Phillips, S. J. et al. (2006). Maximum entropy modeling of species geographic distributions. Ecological Modelling 190, 231-259
}
\keyword{datasets}

