% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.maxnet.R
\name{cv.maxnet}
\alias{cv.maxnet}
\title{Cross-validation for maxnet}
\usage{
cv.maxnet(
  p,
  data,
  f = maxnet.formula(p, data),
  regmult = 1,
  regfun = maxnet.default.regularization,
  addsamplestobackground = T,
  nfolds = 10
)
}
\arguments{
\item{p}{numeric, a vector of 1 (for presence) or 0 (for background)}

\item{data}{a matrix or data.frame of predictor variables}

\item{f}{formula, determines the features to be used}

\item{regmult}{numeric, a constant to adjust regularization}

\item{regfun}{function, computes regularization constant for each feature
By default it uses the \code{\link[=maxnet.default.regularization]{maxnet.default.regularization()}} function to calculate
the base regularization constant for each feature. The output vector is then multiplied
by \code{regmult} to create the vector of \code{penalty.factor}s to pass on to the \code{\link[=glmnet]{glmnet()}} call.}

\item{addsamplestobackground}{logical, if TRUE then add to the background any
presence sample that is not already there}

\item{nfolds}{integer, number of folds - default is 10. Allowable range of values
is 3 to length(p), however, a large number of folds is not recommended for large datasets.}
}
\value{
returns an object of class \code{cv.maxnet}, which is a list
consisting of a cv.glmnet model with the following elements added:
\itemize{
\item \code{betas} nonzero coefficients of the fitted model
\item \code{alpha} constant offset making the exponential model sum to one over the background data
\item \code{entropy} entropy of the exponential model
\item \code{penalty.factor} the regularization constants used for each feature
\item \code{featuremins} minimum of each feature, to be used for clamping
\item \code{featuremaxs} maximum of each feature, to be used for clamping
\item \code{varmin} minimum of each predictor, to be used for clamping
\item \code{varmax} maximum of each predictor, to be used for clamping
\item \code{samplemeans} mean of each predictor over samples (majority for factors)
\item \code{levels} levels of each predictor that is a factor
\item \code{cmx} confusion matrix
\item \code{AUC} Area under the Receiver Operator Characteristic (ROC) curve
\item \code{PCC} Percent Correctly Classified
}
}
\description{
Runs a k-fold cross-validation set for maxnet and plots the resulting ROC curve
}
\seealso{
\code{\link[=cv.glmnet]{cv.glmnet()}}
}
