\name{maxnet-package}
\alias{maxnet-package}
\docType{package}
\title{Maxent over glmnet}
\description{
Maxent species distribution modeling using glmnet for model fitting
}
\details{
\tabular{ll}{
Package: \tab maxnet\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2013-06-06\cr
License: \tab To be determined\cr
}

Create Maxent models for species distributions from presence and background data, using the glmnet package to do the model fitting.  By default, feature sets and regularization are the same as the Maxent java application.
}
\author{
Steven Phillips <phillips@research.att.com>
}
\references{
  Phillips & Dudik
  Fithian & Hastie
  Glmnet
}
\keyword{ package }

