% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxnet.coefficients.R
\name{maxnet.coefficients}
\alias{maxnet.coefficients}
\title{A table of 'features' and lambdas}
\usage{
maxnet.coefficients(envnames, object)
}
\arguments{
\item{envnames}{character vector of model input names. This is required because the maxnet model object only keeps non-zero coefficients, and so may drop original inputs.}

\item{object}{model object of type \code{maxnet}.}
}
\value{
a data.frame or data.table (if \pkg{data.table} installed) with fields \code{input, minval, maxval, penalty, coeff, abscoef}, sorted descending by absolute value of coefficient.
}
\description{
Similar to what was produced in the java Maxent \code{lambdas} file. Note that at this time, hinge features display min/max values in the name, not in the output \code{minval} and \code{maxval} columns.
}
\author{
Michelle Fink
}
