% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxnet.varImp.R
\name{maxnet.varImp}
\alias{maxnet.varImp}
\title{Variable importance for a maxnet model}
\usage{
maxnet.varImp(modobj, resp_table, numReplicates = 5)
}
\arguments{
\item{modobj}{Object of class maxnet}

\item{resp_table}{Data.frame. Response table. First column is assumed to be 'response'}

\item{numReplicates}{Integer. Number of permutations performed to compute importance.}
}
\value{
A named vector of percent importance scores for the variables present in the maxnet model object sorted from highest to lowest.
}
\description{
Computes variable importance scores for all variables present in a maxnet model object. Scores are indicative and should be interpreted with caution as variable interaction is not considered in the computation of the scores.
}
\details{
{
The method used to compute variable importance follows that used in R packages \pkg{biomod2} and \pkg{ecospat}. First, model predictions are made for each row of the environmental data table (or 'response table') \emph{resp_table}.

For each variable in the maxnet model object, values for the variable are permuted between rows and a model prediction made for each row using the permuted or shuffled data table. The permutation is performed \emph{numReplicates} times for each variable.

At each permutation, a Pearson correlation is computed between reference predictions and the predicted values from the shuffled table. The importance score is 1 - correlation coefficient.

A vector of mean scores for each variable expressed as a percentage of the sum of all mean scores is returned.
}
}
\references{
based on peterbat1/fitMaxnet varImportance.R \link{https://rdrr.io/github/peterbat1/fitMaxnet/src/R/varImportance.R}
}
