% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.maxnet.R
\name{predict.maxnet}
\alias{predict.maxnet}
\title{Predict using a maxnet model}
\usage{
\method{predict}{maxnet}(
  object,
  newdata,
  clamp = T,
  type = c("link", "exponential", "cloglog", "logistic"),
  ...
)
}
\arguments{
\item{object}{an object of class "maxnet", i.e., a fitted model.}

\item{newdata}{values of predictor variables to predict to, possibly
matrix, data.frame, \code{SpatRaster} or \code{stars} object.}

\item{clamp}{logical, f true, predictors and features are restricted to the range seen during model training.}

\item{type}{character, type of response required. Using \code{lp} for the linear predictor
and \code{entropy} for the entropy of the exponential model over the background data,
the values returned are determined by the value of \code{type}.
\itemize{
\item{"link"}{yields \code{lp}}
\item{"exponential"}{yields \code{exp(lp)}}
\item{"cloglog"}{yields \code{1-exp(-exp(entropy+lp))}}
\item{"logistic"}{yields \code{1/(1+exp(-entropy-lp))}}
}}

\item{...}{not used}
}
\value{
vector with predicted values (one per input row), \code{SpatRaster} or \code{stars} object of predicted values
}
\description{
Prediction can be on a spatial raster or vector space.
}
