% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.plot.R
\name{response.plot}
\alias{response.plot}
\title{Compute and plot a single response variable}
\usage{
response.plot(
  mod,
  v,
  type,
  mm = mod$samplemeans,
  min = mod$varmin[v],
  max = mod$varmax[v],
  levels = unlist(mod$levels[v]),
  plot = T,
  xlab = v,
  ylab = tools::toTitleCase(type),
  N = 100,
  ...
)
}
\arguments{
\item{mod}{a fitted model, must be of type maxnet if default values used for other arguments.}

\item{v}{charvacter, name of variable to be plotted.}

\item{type}{character, type of response to plot on y-axis.}

\item{mm}{numeric, sample means (or majorities for factors) for predictors; predictors other than v are given these values.}

\item{min}{numeric, minimum value of v; determines range of x-axis}

\item{max}{numeric, maximum value of v; determines range of x-axis}

\item{levels}{vector, if v is a factor, determines levels to be plotted}

\item{plot}{logical, if \code{TRUE} render the plot (or barplot) if \code{FALSE}
then compute the response and return a data.frame}

\item{xlab}{character, label for x-axis}

\item{ylab}{character, label for y-axis}

\item{N}{numeric, the number of intervals over which to sample the response}

\item{...}{other argument passed to plot or barplot}
}
\value{
if \code{plot} is \code{FALSE} a data frame
that contains variable and response columns otherwise \code{NULL} invisibly
}
\description{
Compute and plot a single response variable
}
\seealso{
\code{\link{plot.maxnet}}
}
