% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxnet-package.R
\docType{package}
\name{maxnet-package}
\alias{maxnet-package}
\alias{_PACKAGE}
\alias{maxnet}
\alias{maxent}
\title{Maxent over glmnet}
\description{
Procedures to fit species distributions models from occurrence records and environmental variables, using 'glmnet' for model fitting. Model structure is the same as for the 'Maxent' Java package, version 3.4.0, with the same feature types and regularization options.  See the 'Maxent' website \url{http://biodiversityinformatics.amnh.org/open_source/maxent} for more details.
}
\references{
Phillips, S. and Dudik, M. 2008. Modeling of species distributions with MaxEnt: new extensions and a comprehensive evaluation. Ecography 31:161.
Fithian, W. and Hastie, T. 2012. Finite-sample equivalence of several statistical models for presence-only data. \url{http://arxiv.org/abs/1207.6950v1}.
}
\seealso{
\link{url(https://github.com/mrmaxent/maxnet)} for source code, and \code{\link[=glmnet]{glmnet()}} for base functionality.
}
\author{
Steve Phillips
}
\keyword{internals}
