\name{maxnet-package}
\alias{maxnet-package}
\docType{package}
\title{Maxent over glmnet}
\description{
Maxent species distribution modeling using glmnet for model fitting
}
\details{
\tabular{ll}{
Package: \tab maxnet\cr
Type: \tab Package\cr
Date: \tab 2021-07-08\cr
License: \tab MIT + file LICENSE\cr
}

Create Maxent models for species distributions from presence and background data, using the glmnet package to do the model fitting.  By default, feature sets and regularization are the same as the Maxent java application.
}
\author{
Steven Phillips <phillips@research.att.com>, Author, maintainer

Michelle M. Fink <michelle.fink@colostate.edu>, Contributor
}
\references{
  Phillips, S. and Dudik, M. 2008. Modeling of species distributions
with MaxEnt: new extensions and a comprehensive evaluation. Ecography 31:161.
  
  Fithian, W. and Hastie, T. 2012. Finite-sample equivalence of several
statistical models for presence-only data. <http://arxiv.org/abs/1207.6950v1>.

  package \pkg{glmnet}
}
\keyword{ package }

