% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxnet-package.R
\docType{package}
\name{maxnet-package}
\alias{maxnet-package}
\alias{_PACKAGE}
\title{Maxent over glmnet}
\description{
Procedures to fit species distributions models from occurrence records and environmental variables, using 'glmnet' for model fitting. Model structure is the same as for the 'Maxent' Java package, version 3.4.0, with the same feature types and regularization options.  See the 'Maxent' website \url{http://biodiversityinformatics.amnh.org/open_source/maxent} for more details.
}
\details{
\tabular{ll}{
Package: \tab maxnet\cr
Type: \tab Package\cr
Date: \tab 2021-07-08\cr
License: \tab MIT + file LICENSE\cr
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mrmaxent/maxnet}
}

}
\author{
Steven Phillips <phillips@research.att.com>, Author, maintainer

Michelle M. Fink <michelle.fink@colostate.edu>, Contributor
}
\references{
  Phillips, S. and Dudik, M. 2008. Modeling of species distributions
with MaxEnt: new extensions and a comprehensive evaluation. Ecography 31:161.
  
  Fithian, W. and Hastie, T. 2012. Finite-sample equivalence of several
statistical models for presence-only data. <http://arxiv.org/abs/1207.6950v1>.

  package \pkg{glmnet}
}
\keyword{ package }
