% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxnet-package.R
\docType{package}
\name{maxnet-package}
\alias{maxnet-package}
\alias{_PACKAGE}
\title{Maxent over glmnet}
\description{
Procedures to fit species distributions models from occurrence records and environmental variables, using 'glmnet' for model fitting. Model structure is the same as for the 'Maxent' Java package, version 3.4.0, with the same feature types and regularization options.  See the 'Maxent' website \url{http://biodiversityinformatics.amnh.org/open_source/maxent} for more details.
}
\references{
Phillips & Dudik, Fithian & Hastie, glmnet
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mrmaxent/maxnet}
}

}
\author{
Steve Phillips
}
\keyword{internals}
